+++ src/Enums/HelperStringEncodingEnum.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Atlcom\Enums;
+
+enum HelperStringEncodingEnum: string
+{
+    case Utf8 = 'UTF-8';
+}